refactor: separate clusterize_initializer from initializer, separate cluster_match_strategy to 3 files - utils, score_functions and cluster_match_algorithms
