+++ rework_pysatl_mpest/initializers/clusterize/cluster_match_algorithms.py
@@ -6,154 +6,29 @@ __author__ = "Viktor Khanukaev"
 __copyright__ = "Copyright (c) 2025 PySATL project"
 __license__ = "SPDX-License-Identifier: MIT"
 
-from copy import copy
 from itertools import permutations
-from typing import Any, Callable, TypedDict
+from typing import Callable
 
 import numpy as np
 from scipy.optimize import linear_sum_assignment
 
-from ..core import MixtureModel
-from ..distributions.continuous_dist import ContinuousDistribution
-from ..optimizers import Optimizer, ScipyNelderMead
+from ...core import MixtureModel
+from ...distributions.continuous_dist import ContinuousDistribution
+from ...optimizers import Optimizer, ScipyNelderMead
+from .score_functions import (
+    _calculate_component_aic,
+    _calculate_component_log_likelihood,
+    _calculate_mixture_aic,
+    _calculate_mixture_log_likelihood,
+)
 from .strategies import MatchingMethod, ScoringMethod
-
-MatchingResult = tuple[list[ContinuousDistribution], list[dict[str, float]], list[float]]
-Context = dict[str, Any]
-ScoreFunc = Callable[[MixtureModel | ContinuousDistribution, np.ndarray, np.ndarray | None], float]
-
-
-class FitResult(TypedDict):
-    """A TypedDict to represent the result of a single model-cluster fit.
-
-    Attributes
-    ----------
-    model : ContinuousDistribution
-        The distribution instance with estimated parameters.
-    params : dict[str, float]
-        The estimated parameters (e.g., {'loc': 0.5, 'scale': 1.0}).
-    score : float
-        The goodness-of-fit score (e.g., AIC or negative log-likelihood).
-    weight : float
-        The calculated weight of this component in the mixture.
-    """
-
-    model: ContinuousDistribution
-    params: dict[str, float]
-    score: float
-    weight: float
-
-
-def _validate_clusters_distributions(
-    H: np.ndarray, models_count: int, estimation_strategies_count: int, min_samples: int
-) -> tuple[list[int], list[float]]:
-    """Validates cluster weights and alignment with models.
-
-    Ensures that the responsibility matrix H sums to 1, matches dimensions with
-    provided models, and filters out clusters with insufficient samples.
-    """
-    if not np.allclose(np.sum(H, axis=1), 1, atol=1e-10):
-        raise ValueError("Sum of H matrix weights must be equal to 1")
-
-    n_clusters = H.shape[1]
-
-    if estimation_strategies_count != models_count:
-        raise ValueError("Number of estimation functions must match number of models")
-
-    cluster_weights: list[float] = np.sum(H, axis=0)
-
-    valid_clusters = [k for k in range(n_clusters) if cluster_weights[k] >= min_samples]
-    if len(valid_clusters) != models_count:
-        return [], []
-    return valid_clusters, cluster_weights
-
-
-def _estimate_and_score_component(
-    model: ContinuousDistribution,
-    estimation_func: Callable,
-    score_func: ScoreFunc,
-    X: np.ndarray,
-    H_k: np.ndarray,
-    optimizer: Optimizer,
-) -> FitResult:
-    """Estimates parameters for a model-cluster pair and computes its score.
-
-    Creates a copy of the model, estimates parameters using the provided strategy
-    on the weighted data (H_k), and calculates the fit score.
-    """
-    temp_model = copy(model)
-    new_params: dict[str, float] = estimation_func(temp_model, X, H_k, optimizer)
-    temp_model.set_params_from_vector(list(new_params.keys()), list(new_params.values()))
-
-    score = score_func(temp_model, X, H_k)
-    weight = float(np.sum(H_k) / len(X))
-
-    return {"model": temp_model, "params": new_params, "score": score, "weight": weight}
-
-
-def _calculate_component_log_likelihood(model: ContinuousDistribution, X: np.ndarray, H_k: np.ndarray) -> float:
-    """Calculates the weighted log-likelihood for a single component"""
-    log_probs = np.clip(model.lpdf(X), -1e9, -1e-9)
-    return np.sum(H_k * log_probs)
-
-
-def _calculate_mixture_log_likelihood(model: MixtureModel, X: np.ndarray) -> float:
-    """Calculates the total log-likelihood for a mixture model"""
-    return np.sum(model.loglikelihood(X))
-
-
-def _calculate_component_aic(model: ContinuousDistribution, X: np.ndarray, H_k: np.ndarray) -> float:
-    """Calculates AIC for a single component based on weighted log-likelihood"""
-    weighted_log_likelihood = _calculate_component_log_likelihood(model, X, H_k)
-    k_params = len(model.params)
-    return 2 * k_params - 2 * weighted_log_likelihood
-
-
-def _calculate_mixture_aic(model: MixtureModel, X: np.ndarray) -> float:
-    """Calculates AIC for the entire mixture model"""
-    log_likelihood = _calculate_mixture_log_likelihood(model, X)
-    k_params = sum(len(dist.params) for dist in model.components)
-
-    k_params += model.n_components - 1
-
-    return 2 * k_params - 2 * log_likelihood
-
-
-def _precompute_fits(context: Context) -> list[list[FitResult]]:
-    """Pre-computes and caches parameter estimates and scores for all pairs.
-
-    Iterates through every distribution model and every valid cluster to estimate
-    parameters and calculate the goodness-of-fit score. This results in a cost
-    matrix used by global optimization strategies (Hungarian and Permutations).
-
-    Parameters
-    ----------
-    context : Context
-        The execution context containing models, data (X), weights (H),
-        optimizers, and scoring functions.
-
-    Returns
-    -------
-    list[list[FitResult]]
-        A matrix (list of lists) where entry [i][j] contains the fit result
-        (params, score, weight) for the i-th model and the j-th cluster.
-    """
-    models = context["models"]
-    estimation_strategies = context["estimation_strategies"]
-    valid_clusters = context["valid_clusters"]
-    score_func_component = context["score_func_component"]
-    X, H, optimizer = context["X"], context["H"], context["optimizer"]
-
-    cached_fits: list[list[FitResult]] = []
-
-    for model, est_func in zip(models, estimation_strategies):
-        row: list[FitResult] = []
-        for k in valid_clusters:
-            fit_result = _estimate_and_score_component(model, est_func, score_func_component, X, H[:, k], optimizer)
-            row.append(fit_result)
-        cached_fits.append(row)
-
-    return cached_fits
+from .utils import (
+    Context,
+    MatchingResult,
+    _estimate_and_score_component,
+    _precompute_fits,
+    _validate_clusters_distributions,
+)
 
 
 def _match_greedy(context: Context) -> MatchingResult:
