+++ rework_pysatl_mpest/initializers/clusterize/strategies.py
